refactor: replace validation functions to individual file
